Unit 2 upload
I kind of jumped into the mockup stage before the thumbnail stage. This
is a series of shots of a scrolling page. The hamburger icon will open
a menu across the top with he same links as those in the homepage
boxes. That menu will be sticky through the site. The colored boxes
will actually be photos.
I’m working on another mockup of the site that is separate pages for
each link. If time allows, I will try to upload that as well. Lastly, I
forgot my social media buttons on the contact page.